% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_qfuncmm.R
\name{fit_qfuncmm}
\alias{fit_qfuncmm}
\title{Fit the QFunC Mixed Effects Model}
\usage{
fit_qfuncmm(
  region1_info,
  region2_info,
  start_parms = NULL,
  NNarray = NULL,
  reorder = TRUE,
  group = TRUE,
  m_seq = c(10, 30),
  max_iter = 40,
  fixed_parms = NULL,
  silent = FALSE,
  st_scale = NULL,
  convtol = 1e-04
)
}
\arguments{
\item{region1_info}{stage 1 info list for region 1}

\item{region2_info}{stage 1 info list for region 2}

\item{start_parms}{\eqn{5} vector of starting values. Computed automatically if unspecified.}

\item{NNarray}{Optionally specified array of nearest neighbor indices,
usually from the output of \code{\link{find_ordered_nn}}. If \code{NULL},
fit_model will compute the nearest neighbors. We recommend that the user
not specify this unless there is a good reason to (e.g. if doing a comparison
study where one wants to control
\code{NNarray} across different approximations).}

\item{reorder}{TRUE/FALSE indicating whether maxmin ordering should be used
(TRUE) or whether no reordering should be done before fitting (FALSE).
If you want
to use a customized reordering, then manually reorder \code{y}, \code{locs},
and \code{X},
and then set \code{reorder} to \code{FALSE}. A random reordering is used
when \code{nrow(locs) > 1e5}.}

\item{group}{TRUE/FALSE for whether to use the grouped version of
the approximation (Guinness, 2018) or not.  The grouped version
is used by default and is always recommended.}

\item{m_seq}{Sequence of values for number of neighbors. By default,
a 10-neighbor
approximation is maximized, then a 30-neighbor approximation is
maximized using the
10 neighbor estimates as starting values.
However, one can specify any sequence
of numbers of neighbors, e.g. \code{m_seq = c(10,30,60,90)}.}

\item{max_iter}{maximum number of Fisher scoring iterations}

\item{fixed_parms}{Indices of covariance parameters you would like to fix
at specific values. If you decide to fix any parameters, you must specify
their values in \code{start_parms}, along with the starting values for
all other parameters. For example, to fix the nugget at zero in
\code{exponential_isotropic}, set \code{fixed_parms} to \code{c(3)}, and set
\code{start_parms} to \code{c(4.7,3.1,0)}. The
last element of \code{start_parms} (the nugget parameter) is set to zero,
while the starting values for the other two parameters are 4.7 and 3.1.}

\item{silent}{TRUE/FALSE for whether to print some information during fitting.}

\item{st_scale}{Scaling for spatial and temporal ranges.}

\item{convtol}{Tolerance for exiting the optimization.
Fisher scoring is stopped
when the dot product between the step and the gradient
is less than \code{convtol}.}
}
\description{
Given a response, set of locations, (optionally) a design matrix,
and a specified covariance function, return the maximum
Vecchia likelihood estimates, obtained with a Fisher scoring algorithm.
}
