% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GpGpQFuncMM.R
\docType{package}
\name{GpGpQFuncMM}
\alias{GpGpQFuncMM}
\alias{GpGpQFuncMM-package}
\title{GpGpQFuncMM: Stage 2 of QFuncMM via Vecchia's Approximation}
\description{
Employs Vecchia's approximation to fit stage 2 of the QFunC Mixed Model. The code is largely based on the `GpGp` package by Joe Guiness. The original approximation is due to Vecchia (1988) \url{http://www.jstor.org/stable/2345768}, and the reordering and grouping methods are from Guinness (2018) \doi{10.1080/00401706.2018.1437476}. Model fitting employs a Fisher scoring algorithm described in Guinness (2019) \doi{10.48550/arXiv.1905.08374}.
}
\details{
The main function of the package is \code{\link{fit_qfuncmm}}.
}
\author{
\strong{Maintainer}: Robin Liu \email{robin28liu@gmail.com}

}
