% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordering_functions.R
\name{order_qfuncmm}
\alias{order_qfuncmm}
\title{Ordering for Qfunc mixed model}
\usage{
order_qfuncmm(locs, space_cols = seq(1L, 3L), time_col = 4L, region_col = 5L)
}
\arguments{
\item{locs}{A matrix of locations. Each row of \code{locs} contains a location, which can
be a point in Euclidean space R^d, a point in space-time R^d x T, 
a longitude and latitude (in degrees) giving a point on the sphere, 
or a longitude, latitude, and time giving a point in the sphere-time domain.}

\item{space_cols}{column indices of spatial dimensions of \code{locs}}

\item{time_col}{column index of time dimension of \code{locs}}

\item{region_col}{column index of region dimension of \code{locs}}
}
\value{
A vector of indices giving the ordering.
}
\description{
First the time dimension is ordered middle-out. For each time point, the voxels are
ordered using maximum minimum distance ordering independently for each region.
Finally, these within-region orderings are interleaved.
}
