% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{test_qfunc_cov}
\alias{test_qfunc_cov}
\title{Compute QFunc Covariance Matrix.}
\usage{
test_qfunc_cov(stage1_parms, covparms, locs)
}
\arguments{
\item{stage1_parms}{A 2 x 5 matrix of stage 1 paramerters.}

\item{covparms}{5-vector of stage1 parameters.}

\item{locs}{A matrix with \code{n} rows and 5 columns.}
}
\description{
Compute QFunc Covariance Matrix.
}
